sqlness_runner
Module util

Constants

PORT_CHECK_INTERVAL 🔒

Check port every 0.1 second.

Functions

check_port

Spin-waiting a socket address is available, or timeout. Returns whether the addr is up.

get_binary_dir

get_case_dir

Get the dir of test cases. This function only works when the runner is run under the project’s dir because it depends on some envs set by cargo.

get_workspace_root

Get the dir that contains workspace manifest (the top-level Cargo.toml).

setup_wal

Start kafka cluster if needed. Config file is conf/kafka-cluster.yml.

sqlness_conf_path

Get the path of sqlness config dir tests/conf.

teardown_wal

Stop kafka cluster if needed. Config file is conf/kafka-cluster.yml.